backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Environment and configuration
python-decouple>=3.8
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI and LLM
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.10

# HTTP requests and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Security and validation
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Logging and observability
structlog>=23.2.0
python-json-logger>=2.0.7

# Redis for caching
redis>=5.0.0
hiredis>=2.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development and utilities
python-dotenv>=1.0.0